feature: recorrido (#72)
* feat: Inicio and Find endpoints

* feature: third endpoint, put

* feat: all set but post tests

* fix: numerous fixes

* feat: recorrido endpoints finished

---------

+++ src/main/java/org/circle8/controller/response/PuntoResponse.java
@@ -4,6 +4,6 @@ import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
 public class PuntoResponse implements ApiResponse {
-	public float latitud;
-	public float longitud;
+	public double latitud;
+	public double longitud;
 }
